Add pudb
Note that this breaks from the pip pudb install slightly in that it always
uses `pudb` as the entry-point, not pudb3 for Python 3. I'm one of the PuDB
developers, and I think that `pudb3` is a bad idea. If you want we can ask
Andreas what he thinks as well.