add datreant recipes.
This includes two recipes for datreant. One is the empty namespace
package 'datreant' and the other contains the actual core pacakge
for datreant. The config for datreant itself is to allow to install
a simple namespace package. For 'datreant.core' the extra build scripts
are needed to remove the '*-nspkg.pth' file produced by setuptools,
this is to have anaconda-verify pass with the package.