Add tests to bc recipe
On OS X 10.11.5 and Ubuntu 16.04, this recipe builds a bc that fails
on exactly 10 tests. On both platforms, the system bc fails on the same
tests with the same values. This seems par for the course for bc:

http://www.linuxfromscratch.org/blfs/view/7.4/general/bc.html

The tests added here simply verify that that only those 10 tests have
failed.