During conversion, skip feedstocks that fail CI registration.
Ideally we would run the full conversion process on each recipe
independently. That way we are guaranteed to bank our progress with each
one. So if something goes wrong with the conversion of one recipe, it
doesn't cost us our progress. However, this appears not to work due to
some issues with doing the full Travis CI sync. Still that does not
lessen the need for some ability to do partial conversions.

The biggest driver for this is that CI registration can be flaky
sometimes. While the ideal solution would be to figure out the
underlying CI issue and solve the problem or at least workaround it, in
practice this is a challenging ask. There are many reasons for this, but
they ultimately hinge on the fact that we don't know how the CIs are
doing things under the hood and so we can't be sure what corner cases we
might be hitting. The other is that we are only one group of users on
these very popular services. So we may need to wait in the queue for
awhile until we actually get time spent on our issue. In the interim,
these CI issues persist. Improved resilience on our part will help limit
our vulnerability.

In order to solve this issue, we implement the simplest way to bank our
progress during feedstock conversion. Basically if a feedstock fails CI
registration, we note the failure and skip over it. This way if it is
something funny with that feedstock we can still get the other ones out
so we can look at that one more closely. However, if it is something
else like a limitation on how many we can process at a time, we can at
least get a few out so that we will have less to work through next time.
Once we have tried to convert and remove everything, we then mark the
build as failed. This way we are alerted to the fact there is a problem
still, but we have cut down the amount of work that we have to do.
Perhaps a simple restart will be sufficient.