Adjusts purge merged recipes trick.
After a quick look it seems all the CIs clone `master`. This includes
whether they are building PRs or are running on `master`. So there is no
real need for fetching `master`. Also, this should avoid the occasional
discontinuity where the fetched `master` differs from the one previously
cloned. Resolving these discontinuities will ensure all recipes that
need to be purged are. Note this guarantees a fast finish when a CI
builds on `master`.