create_feedstocks: Enable autostash with rebase.
Apparently our `git` doesn't not expose the `--autostash` flag for `git
pull`. Appears that flag is not added until `git` version `2.9.0`. Here
is the commit (
https://github.com/git/git/commit/f66398eb57c627169429f47bbe4d943d2c975959
) that appears to add it.

However, we can work around this minor limitation by simply allowing
rebases to do an autostash. Here we enable such a configuration to keep
life simple. Tried this out locally with our `git` and verified that it
worked.

[ci skip]