During conversion, resolve new file conflicts due to updating master.
It is possible when popping the stash after updating to match the remote
`master` that there are some new files included in upstream that are now
marked as untracked (as they weren't included when the stash was
created). An example of this is a new recipe that got merged into
`master` before we started conversion.

Obviously we don't want them marked as untracked. So we go ahead add
them to the index. As they are already tracked and in the index as is
this is actually a no-op. However, it does resolve the conflict and let
us proceed with removal of the converted recipes.