Consolidate selector and compare exact versions.
There were 3 `skip`s in this recipe with different selectors and they
were not mutually exclusive so one could end up with more than one
`skip` in a recipe, which is not valid. This consolidates the `skip`s
under one selector. Also the `<` comparison for `np` was raised as
problematic. So this adjust that to an `==` comparison, which is known
to work.

cc @dolfim