Use external fast finish CI script
Use the fast finish CI script from `conda-forge-build-setup` to cancel
old PR builds on the CIs. As all of the CIs have some version of Python
either 2 or 3 and the script supports both, this works just fine as soon
as the CI boots up. As it is external, it is a simple matter for all of
the CIs to `curl` the script and run it. All updates made to the script
will be immediately available for any CI that uses the script.

[skip lint]