Travis: Download Miniconda install script directly
Instead of making use of Obvious-CI's bootstrap script to install
Miniconda, simply download this script ourselves. This is good for a
number of reasons including Obvious-CI is being phased out throughout
conda-forge, Travis CI is the only CI still making use of the bootstrap
script to install Miniconda, bugs can be more easily fixed in
conda-forge than Obvious-CI as there are more hands on deck, and
additional optimizations to the setup can be made like caching the
install script. For all of these reasons and more, we switch over to
directly downloading and installing Miniconda on Travis CI.

[skip lint]
[skip appveyor]