X.org libICE: fix build on 64-bit Windows
I guess on 32-bit Windows, `long` and `time_t` are identical types, but
they're not on 64-bit Windows, so `iceauth.c` fails to compile since it
includes a random old-school manual declaration of the `time()` function. The
file unconditionally includes `time.h` so the declaration should be totally
unneccessary — my guess is that code path hasn't been compiled for a looong
time (it's in an #ifdef).