poppler: work around to get this building on OSX
The `lcms` library that we were using is one of the ones from Anaconda
`defaults` that has a broken "soname" parameter without an `@rpath/` prefix,
which makes it challenging to build against with the way that OSX library
finding works. The dependency is optional so it's easiest to just drop it for
now.

Now that `gobject-introspection` has been fixed to properly matrix across
Python versions, we run the build on Python 3.6 just to make sure that we're
working with the latest versions.