Add reentry recipe
Adding a recipe for the reentry python package.

Currently, the tests don't pass when using the archive
provided on PyPI (since a file is missing from the .tar).
I have raised an issue (and a pull request to solve it).

For the moment, I rely on a direct checkout of the github repository.