Force removal of recipes after conversion
In case for some reason a converted recipe cannot be removed, use force
to ensure it is removed from the staged-recipes repo. This may occur if
there are formatting changes due to `.gitattributes` for instance. As
the plan is to remove the recipe completely anyways, using force should
not be a problem.

[skip ci]