elfx86exts: skip build on OSX
The dependent Rust crate "capstone-sys" currently doesn't build on OSX. This
looks to be a superficial problem but we can fix it later. One might want to
disassemble ELF-format binaries on OSX but seeing as that OS uses a different
format for binaries it's definitely a lower priority to make this tool
available on that platform.