Drop webpage CI trigger
This is no longer needed as the feedstock listing is now generated and
stored in the feedstocks repo using a webservice. This will be
accurately updated using the feedstocks repo's content any time there is
a push to staged-recipes. In practice, this means that listing will be
updated any time staged-recipes clears out some or all recipes during
feedstock conversion as they will already have been added to the
feedstocks repo as part of conversion. By switching to the webservice,
this will cutdown rate limiting problems that the old strategy webpage
based strategy was bound to encounter. Also this will be more reliable
during partial conversions.

[ci skip]