Ensure that `build_artifacts` has a `noarch` index
The `noarch` index doesn't seem to be generated by `conda-build` in this
case as we are using a non-standard API. However not having `noarch`
seems to cause `conda` to fail in this case. To fix this issue, go ahead
and create a dummy `noarch` index so this appears to be a valid channel.
This seems to satisfy `conda` and allows the build to proceed.

[skip lint]