Correct file hash
The source tarball has *not* changed. What has changed is the Content-Encoding header
sent by the webserver. Previously, this server sent the x-gzip encoding. conda-build
listened to this header and transparanetly decompressed the file. The previous hash
was the hash of the .tar file after un-gzipping. The new hash is the hash of the .tar.gz
file as it should be.