Package license information with the CDTs
This includes the copyright notice taken from the README explaining the
licensing of numactl and libnuma. Also have included the generic GPL 2.0
and and LGPL 2.1 license files, which apply to each part respectively.
This matches what the author's recommend we do to acquire a license
file. There's a special section about numademo, which doesn't seem
relevant for our use case; so, have ignored it.