fix conda.build warning
conda.build produces this warning on linux:

```
WARNING (libvips,lib/libvips-cpp.so.42.11.0): Needed DSO lib/libxml2.so.2 found in ['libxml2']
WARNING (libvips,lib/libvips-cpp.so.42.11.0): .. but ['libxml2'] not in reqs/run, (i.e. it is overlinking) (likely) or a missing dependency (less likely)
```

Even though libxml2 is not a direct dependency of libvips. Perhaps a
package we depend on uses libxml2 but does not declare it?

This commit adds libxml2 as a run dependency to fix this warning.