add recipe for openslide
This is recipe is copy-pasted from:

https://github.com/bioconda/bioconda-recipes/tree/master/recipes/openslide

I'm trying to add a conda package for libvips, and as well as being a
general-purpose image processing library, libvips can load images with
openslide.

Openslide is currently in bioconda, so I can't list it as a libvips dep
in conda-forge. I'm therefore opening this PR as a first step in trying to
move openslide from bioconda to conda-forge.