Fixed qt version, and add test for CLI program.
There is a problem with qt version 5.12.*, which causes
a dependency issue (lib64/libEGL.so can't be found). This
does not occur with qt version 5.9.*, so fixing the qt version
for the moment.

Also added a test, using the SI100.XYZ input file. I've not
commented out the "prismatic --help" test, as this returns a
non-zero error code (fooling conda build into thinking it has
failed).