Add Documenteer (#10614)
* Add sphinx-prompt recipe

* Add sphinx-jinja recipe

* Add lsst-sphinx-bootstrap-theme recipe

* Add lsst-dd-rtd-theme recipe

* Add documenteer recipe

* Add import test for sphinx-jinja

* Add setuptools as a run dependency

These packages use pkg_resources, so they need setuptools in their
runtime.

* Add sphinx dependency for sphinx-jinja

* Add run_test.py for sphinx-prompt

This test script establishes that the Sphinx is importable and that it
has a setup() function.

* Update to sphinx-jinja 1.1.1

This new version of sphinx-jinja includes the LICENSE file in its
distribution.

* Set python version properly

This does away with the build: skip [py27] approach and properly
establishes versions of Python needed for run and host.

All of these packages are also pure python, so we use build.noarch:
python.

Extra notes:

- documenteer 0.5.6 requires Python 3.6+ at runtime, hence that
  constraint
- sphinx-jinja is both Python 2.7 and 3+ compatible.

* Change lsst_dd_rtd_theme package name

It's "lsst-dd_rtd_theme" on PyPI, but the Python package name is
"lsst_dd_rtd_theme". That the package name is not "lsst-dd-rtd-theme" in
the setuptools metadata, to match the PyPI name, is probably an error
from our earlier by-hand releases.

* Use lsst-dd-rtd-theme 0.2.3

In this new release of lsst-dd-rtd-theme we fixed up the packaging
metadata on PyPI. The package name matches the PyPI name and is
lsst-dd-rtd-theme. The Python module is lsst_dd_rtd_theme.

* Update lsst-dd-rtd-theme pin

Use the newer release of lsst-dd-rtd-theme that's suitable for Conda
distribution.

* Update sphinx-jinja pin to 1.1.1

* Add lsst-documenteer-pipelines -technote recipes

These two meta packages provide the dependencies, pins, and command-line
tools necessary for specific LSST documentation projects:
https://pipelines.lsst.io and technical notes, respectively.

This leaves lsst-documenteer as as a core package for the Sphinx
extensions provided by documenteer itself. Note that the
lsst-documenteer core package gains a specific dependency on pybtex
since it no longer has a dependency on sphinxcontrib-bibtex.

* Move all entry_points back to lsst-documenteer

The windows build required that lsst-documenteer package declared all
of its entrypoints, but we'll only "test" the relevant command-line
programs in the application-specific metapackages.

* Drop pip installs from meta packages

* Collect meta packages as outputs

Now lsst-documenteer-technote and lsst-documenteer-pipelines are defined
within the same recipe as documenteer, but succeed in keeping the
meta package dependencies contained.

* Fix copy-pasted fancy quotes

* Drop host sections from metapackages