Add cogdl
I was a bit verbose/explicit with dependencies.  I base them on what is imported by modules within the package.

`dgl` could be added as a dependency if it's added to conda-forge.  See dmlc/dgl#1855 and conda-forge/staged-recipes#12522

Even though this is a pure Python package, I run a few of the fastest tests that exercise different files, because
this package has some heavy dependencies (pytorch, pytorch_geometric, etc) and I want that warm fuzzy feeling :)

I get the tarball from github instead of PyPI, because the PyPI tarball does not (yet) contain the license file or the tests.
Currently, the github tarball is not that much larger than from PyPI.