Drop test finding lib through pkg-config.
Since this needed a run dependency as a workaround to get pkg-config
installed for the testing phase, that means packages building against
libhackrf would get pkg-config in their host environment when it really
should be in the build environment. We don't really need this test, so
just drop it for simplicity.