move openssh to test requirements
@carterbox recommended:
openssh is only required because these packages are being built inside a container
It's not actually a direct dependency of your package.
Only direct dependencies of your package go in the requirements section.
Packages that are only required for the tests go in the test/requires section,
not the requirements/run section.