fix: flit build and bump stdlibs version
In order to stop the flit-based metadata generation from breaking, we are
patching the package source. The change that is introduced simply moves
the package version into __init__.py. This prevents Flit from doing an
import of the module in order to derive version and module docstring.

Furthermore we bump the stdlibs version to 2022.10.9.